ColumnSchema: fix setDataType for timestamp columns
When changing database type of a column, default value remains the same.
That is ok for literal default values.

However, for specific default value CURRENT_TIMESTAMP this is a problems,
because this default value can't be used for column types other than timestamp.

To fix this issue, we reset this specific default when changing database type